Structured the repo and updated requirements.txt
