+++ .idea/inspectionProfiles/Project_Default.xml
@@ -1,0 +1,20 @@
+<component name="InspectionProjectProfileManager">
+  <profile version="1.0">
+    <option name="myName" value="Project Default" />
+    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
+      <option name="ignoredPackages">
+        <value>
+          <list size="7">
+            <item index="0" class="java.lang.String" itemvalue="langchain-community" />
+            <item index="1" class="java.lang.String" itemvalue="faiss-cpu" />
+            <item index="2" class="java.lang.String" itemvalue="langchain" />
+            <item index="3" class="java.lang.String" itemvalue="langchain-google-genai" />
+            <item index="4" class="java.lang.String" itemvalue="PyPDF2" />
+            <item index="5" class="java.lang.String" itemvalue="pypdf2" />
+            <item index="6" class="java.lang.String" itemvalue="gemini" />
+          </list>
+        </value>
+      </option>
+    </inspection_tool>
+  </profile>
+</component>
