+++ requirements.txt
@@ -1,1 +1,3 @@
-gemini
+setuptools
+google-generativeai
+requests
